[feature] Check user's setted avatar icon extension           format is jpg or png.

+++ lib/userhooks.php
@@ -24,12 +24,16 @@ class UserHooks {
     public function register() {
         $loginRecord = function($user) {            
             $view = new \OC\Files\View('/'.$user->getUID());
-            if (!$view->file_exists('avatar.png')){
-                $view->file_put_contents('avatar.png',file_get_contents(\OC_App::getAppPath('setting_default_avatar').'/img/avatar.png'));
-            }
-            
-        };
+            if (!$view->file_exists('avatar.png') and !$view->file_exists('avatar.jpg')){
+                if (file_exists(\OC_App::getAppPath('setting_default_avatar').'/img/avatar.png')){
 
+                    $view->file_put_contents('avatar.png',file_get_contents(\OC_App::getAppPath('setting_default_avatar').'/img/avatar.png'));
+                 }
+                else if(file_exists(\OC_App::getAppPath('setting_default_avatar').'/img/avatar.jpg')){
+                     $view->file_put_contents('avatar.jpg',file_get_contents(\OC_App::getAppPath('setting_default_avatar').'/img/avatar.jpg'));
+                     }
+            }
+        };
         $this->userManager->listen('\OC\User', 'postLogin', $loginRecord);
 
         }
